# mobile/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "每日減脂計畫 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/users/me/profile": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Read Profile",
        "operationId": "read_profile_users_me_profile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileWithTargetsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Replace Profile",
        "operationId": "replace_profile_users_me_profile_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileWithTargetsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update Profile",
        "operationId": "update_profile_users_me_profile_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileWithTargetsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/targets/preview": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Preview Targets",
        "description": "The profile-setup screen shows targets as the user types. Nothing is stored.",
        "operationId": "preview_targets_users_me_targets_preview_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/reminders": {
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update Reminders",
        "operationId": "update_reminders_users_me_reminders_patch",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemindersPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete Me",
        "operationId": "delete_me_users_me_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileIn": {
        "properties": {
          "sex": {
            "type": "string",
            "enum": [
              "f",
              "m"
            ],
            "title": "Sex"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "height_cm": {
            "type": "number",
            "maximum": 250.0,
            "minimum": 100.0,
            "title": "Height Cm"
          },
          "weight_kg": {
            "type": "number",
            "maximum": 300.0,
            "minimum": 30.0,
            "title": "Weight Kg"
          },
          "activity_level": {
            "type": "string",
            "enum": [
              "sedentary",
              "light",
              "moderate",
              "active"
            ],
            "title": "Activity Level"
          },
          "deficit_pct": {
            "type": "integer",
            "enum": [
              10,
              12,
              15,
              20
            ],
            "title": "Deficit Pct"
          },
          "workout_time": {
            "type": "string",
            "enum": [
              "am",
              "pm"
            ],
            "title": "Workout Time",
            "default": "pm"
          },
          "default_location": {
            "type": "string",
            "enum": [
              "home",
              "gym"
            ],
            "title": "Default Location",
            "default": "home"
          },
          "reminder_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{2}:\\d{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Time"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "Australia/Brisbane"
          }
        },
        "type": "object",
        "required": [
          "sex",
          "birth_date",
          "height_cm",
          "weight_kg",
          "activity_level",
          "deficit_pct"
        ],
        "title": "ProfileIn"
      },
      "ProfileOut": {
        "properties": {
          "sex": {
            "type": "string",
            "title": "Sex"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "height_cm": {
            "type": "number",
            "title": "Height Cm"
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg"
          },
          "activity_level": {
            "type": "string",
            "title": "Activity Level"
          },
          "deficit_pct": {
            "type": "integer",
            "title": "Deficit Pct"
          },
          "workout_time": {
            "type": "string",
            "title": "Workout Time"
          },
          "default_location": {
            "type": "string",
            "title": "Default Location"
          },
          "reminder_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Time"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          }
        },
        "type": "object",
        "required": [
          "sex",
          "birth_date",
          "height_cm",
          "weight_kg",
          "activity_level",
          "deficit_pct",
          "workout_time",
          "default_location",
          "reminder_time",
          "timezone"
        ],
        "title": "ProfileOut"
      },
      "ProfilePatch": {
        "properties": {
          "sex": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "f",
                  "m"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 100.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 300.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "activity_level": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "sedentary",
                  "light",
                  "moderate",
                  "active"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Level"
          },
          "deficit_pct": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  10,
                  12,
                  15,
                  20
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Deficit Pct"
          },
          "workout_time": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "am",
                  "pm"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Workout Time"
          },
          "default_location": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "home",
                  "gym"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Location"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "type": "object",
        "title": "ProfilePatch"
      },
      "ProfileWithTargetsOut": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ProfileOut"
          },
          "targets": {
            "$ref": "#/components/schemas/TargetsOut"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "targets"
        ],
        "title": "ProfileWithTargetsOut"
      },
...
      "RemindersPatch": {
        "properties": {
          "reminder_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{2}:\\d{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reminder Time"
          }
        },
        "type": "object",
        "title": "RemindersPatch"
      },
...
      "TargetsOut": {
        "properties": {
          "bmr": {
            "type": "integer",
            "title": "Bmr"
          },
          "tdee": {
            "type": "integer",
            "title": "Tdee"
          },
          "base_kcal": {
            "type": "integer",
            "title": "Base Kcal"
          },
          "exercise_kcal": {
            "type": "integer",
            "title": "Exercise Kcal"
          },
          "kcal": {
            "type": "integer",
            "title": "Kcal"
          },
          "protein_g": {
            "type": "integer",
            "title": "Protein G"
          },
          "fat_g": {
            "type": "integer",
            "title": "Fat G"
          },
          "carb_g": {
            "type": "integer",
            "title": "Carb G"
          }
        },
        "type": "object",
        "required": [
          "bmr",
          "tdee",
          "base_kcal",
          "exercise_kcal",
          "kcal",
          "protein_g",
          "fat_g",
          "carb_g"
        ],
        "title": "TargetsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}